ai-worker/stylist/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==1.1.5
langchain-anthropic==1.4.1
pydantic==2.12.5
python-dotenv==1.2.2